feat: refactor monitor
