+++ llm/context/docs_tree.md
@@ -1,0 +1,6 @@
+#!base_path=/home/nicholasgrundl/projects/ragnostic/docs
+docs
+├── 1_System_Requirements.md
+├── 2_Ragnostic_Project_Plan.md
+├── 2a_Document_Ingestion.md
+└── 2b_Semantic_Extraction.md
