+++ src/ragnostic/ingestion/monitor/__init__.py
@@ -1,0 +1,9 @@
+"""Directory monitoring functionality for document ingestion."""
+from .monitor import DirectoryMonitor
+from .schema import MonitorResult, MonitorStatus
+
+__all__ = [
+    "DirectoryMonitor",
+    "MonitorResult",
+    "MonitorStatus",
+]
