+++ tests/ingestion_monitor/__init__.py
@@ -1,0 +1,1 @@
+"""Monitor tests package."""
